Fix for Windows: use path.sep instead of /
I tested this package on Linux and it works great, but it failed to work
on Windows.

After digging a bit, I discovered that the code is using / as path
separator which unfortunately doesn't work on Windows.
This uses [path.sep](https://nodejs.org/api/path.html#path_path_sep) to
make it work cross platform.